Refactor Settings-overview (IOS-14)
